Added Image resizing, Improved performance of getting colors, Moved Compact List to a file, Added BlurView

+++ MangaReader/Networking/Models/Comic.swift
@@ -7,7 +7,7 @@ struct Comic: Codable {
     let country: String
     let status: Int
     let links: Links?
-    let lastChapter: Int
+    let lastChapter: Double
     let chapterCount: Int
     let demographic: Int?
     let hentai: Bool
